refactor: move lightFire and chance methods into separate files
